refactor(bridge): reorganize package structure
리팩토링 : 패키지 이동이 허용된 클래스에 한해, MVC 패턴으로 구조 변경
