+++ src/main/java/bridge/model/Bridge.java
@@ -1,0 +1,4 @@
+package bridge.model;
+
+public class Bridge {
+}
